requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
python-dotenv>=1.0.0
pyodbc>=5.0.1
SQLAlchemy>=2.0.23
typing-extensions>=4.8.0
urllib3>=2.1.0
